Move the ReadyQueue into FuturesUnordered
The existing FuturesUnordered type does basically the exact same thing, so let's
just reuse that instead for now.
